🙀🔫 Checkpoint ./commands/say.js:3080416/148

+++ commands/say.js
@@ -3,7 +3,10 @@ const errors = require("../utils/errors.js");
 
 module.exports.run = async (client, message, args) => {
 
-
+  if(!args [0] || args[0] == "help"){
+    message.reply("Usage: !say <word or sentence>");
+    return;
+  }
   if(!message.member.hasPermission("MANAGE_MESSAGES")) return errors.noPerms(message, "MANAGE_MESSAGES");
   message.delete();
   let botmessage = args.join(" ");
